schedule and exam-schedule backend completed.

+++ App/laravel/app/Http/Controllers/ExamScheduleController.php
@@ -83,8 +83,13 @@ class ExamScheduleController extends Controller
     // ** permanentDelete method for forceDelete
     public function permanentDelete($id)
     {
-        $examSchedule = ExamSchedule::findOrFail($id);
-        $examSchedule->forceDelete();
+        $examSchedule = ExamSchedule::onlyTrashed()->find($id);
+        if(!is_null($examSchedule)){
+            $examSchedule->forceDelete();
+        }
+        else{
+            return response()->json(['error'=>"Trashed Object Not Found"], 404);
+        }
     }
     // ** restore method for restoring softDeletes records
     public function restore($id)
